[SRP] wip: Extraindo duplicação para uma classe renderizadora de Markdown
